add missing bindings and fix compilation
