+++ python/bindings/src/node_attributes.cpp
@@ -37,6 +37,7 @@
 #include <pybind11/chrono.h>
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 #include <spark_dsg/node_attributes.h>
 
 #include <iomanip>
